Switch to a single-threaded implementation
We replace `Arc` with `Rc`, and we replace the global symbol table with
a thread-local `Context`. This still allows us to run our unit tests in
multiple threads, because we can give each thread its own symbol table.

But when we run in a single WASM thread, we no longer need to pay for
the size overhead of `Arc`.